Break out constants file
