+++ frontend/src/etc/constants/base.ts
@@ -1,0 +1,10 @@
+// ** Base ** //
+
+export const inputTypes = {
+  MULTISELECT: "multiselect",
+  CHIPS: "chips",
+  SELECT: "select",
+  DATE: "date",
+  INPUT_TEXT: "inputText",
+  CATEGORIZED_VALUE: "categorizedValue",
+};
