+++ frontend/src/etc/constants/events.ts
@@ -1,16 +1,11 @@
-import { eventRemediationRead } from "./../models/eventRemediation";
+import { eventRemediationRead } from "../../models/eventRemediation";
 import { filterOption } from "@/models/base";
 import { eventPreventionToolRead } from "@/models/eventPreventionTool";
 import { eventVectorRead } from "@/models/eventVector";
 import { nodeThreatRead } from "@/models/nodeThreat";
 import { nodeThreatActorRead } from "@/models/nodeThreatActor";
 import { observableTypeRead } from "@/models/observableType";
-
 import { useAlertDispositionStore } from "@/stores/alertDisposition";
-import { useAlertQueueStore } from "@/stores/alertQueue";
-import { useAlertToolStore } from "@/stores/alertTool";
-import { useAlertToolInstanceStore } from "@/stores/alertToolInstance";
-import { useAlertTypeStore } from "@/stores/alertType";
 import { useEventPreventionToolStore } from "@/stores/eventPreventionTool";
 import { useEventQueueStore } from "@/stores/eventQueue";
 import { useEventRemediationStore } from "@/stores/eventRemediation";
@@ -24,262 +19,7 @@ import { useNodeThreatStore } from "@/stores/nodeThreat";
 import { useNodeThreatActorStore } from "@/stores/nodeThreatActor";
 import { useObservableTypeStore } from "@/stores/observableType";
 import { useUserStore } from "@/stores/user";
-
-// ** Base ** //
-
-export const filterTypes = {
-  MULTISELECT: "multiselect",
-  CHIPS: "chips",
-  SELECT: "select",
-  DATE: "date",
-  INPUT_TEXT: "inputText",
-  CATEGORIZED_VALUE: "categorizedValue",
-};
-
-// ** Alerts ** //
-
-export const alertPropertyTypes: Record<string, string> = {
-  DISPOSITION_FILTER: "disposition",
-  DISPOSITION_USER_FILTER: "dispositionUser",
-  DISPOSITIONED_AFTER_FILTER: "dispositionedAfter",
-  DISPOSITIONED_BEFORE_FILTER: "dispositionedBefore",
-  EVENT_UUID_FILTER: "eventUuid",
-  EVENT_TIME_AFTER_FILTER: "eventTimeAfter",
-  EVENT_TIME_BEFORE_FILTER: "eventTimeBefore",
-  INSERT_TIME_AFTER_FILTER: "insertTimeAfter",
-  INSERT_TIME_BEFORE_FILTER: "insertTimeBefore",
-  NAME_FILTER: "name",
-  OBSERVABLE_FILTER: "observable",
-  OBSERVABLE_TYPES_FILTER: "observableTypes",
-  OBSERVABLE_VALUE_FILTER: "observableValue",
-  OWNER_FILTER: "owner",
-  QUEUE_FILTER: "queue",
-  TAGS_FILTER: "tags",
-  THREAT_ACTOR_FILTER: "threatActor",
-  THREATS_FILTER: "threats",
-  TOOL_FILTER: "tool",
-  TOOL_INSTANCE_FILTER: "toolInstance",
-  TYPE_FILTER: "type",
-};
-
-export const alertFilters: readonly filterOption[] = [
-  {
-    name: alertPropertyTypes.DISPOSITION_FILTER,
-    label: "Disposition",
-    type: filterTypes.SELECT,
-    store: useAlertDispositionStore,
-    optionProperty: "value",
-    valueProperty: "value",
-  },
-  {
-    name: alertPropertyTypes.DISPOSITION_USER_FILTER,
-    label: "Dispositioned By",
-    type: filterTypes.SELECT,
-    store: useUserStore,
-    optionProperty: "displayName",
-    valueProperty: "username",
-  },
-  {
-    name: alertPropertyTypes.DISPOSITIONED_AFTER_FILTER,
-    label: "Dispositioned After",
-    type: filterTypes.DATE,
-    stringRepr: (filter: Date) => {
-      return filter.toISOString();
-    },
-    parseStringRepr: (valueString: string) => {
-      return new Date(valueString);
-    },
-  },
-  {
-    name: alertPropertyTypes.DISPOSITIONED_BEFORE_FILTER,
-    label: "Dispositioned Before",
-    type: filterTypes.DATE,
-    stringRepr: (filter: Date) => {
-      return filter.toISOString();
-    },
-    parseStringRepr: (valueString: string) => {
-      return new Date(valueString);
-    },
-  },
-  {
-    name: alertPropertyTypes.EVENT_TIME_AFTER_FILTER,
-    label: "Event Time After",
-    type: filterTypes.DATE,
-    stringRepr: (filter: Date) => {
-      return filter.toISOString();
-    },
-    parseStringRepr: (valueString: string) => {
-      return new Date(valueString);
-    },
-  },
-  {
-    name: alertPropertyTypes.EVENT_TIME_BEFORE_FILTER,
-    label: "Event Time Before",
-    type: filterTypes.DATE,
-    stringRepr: (filter: Date) => {
-      return filter.toISOString();
-    },
-    parseStringRepr: (valueString: string) => {
-      return new Date(valueString);
-    },
-  },
-  {
-    name: alertPropertyTypes.INSERT_TIME_AFTER_FILTER,
-    label: "Insert Time After",
-    type: filterTypes.DATE,
-    stringRepr: (filter: Date) => {
-      return filter.toISOString();
-    },
-    parseStringRepr: (valueString: string) => {
-      return new Date(valueString);
-    },
-  },
-  {
-    name: alertPropertyTypes.INSERT_TIME_BEFORE_FILTER,
-    label: "Insert Time Before",
-    type: filterTypes.DATE,
-    stringRepr: (filter: Date) => {
-      return filter.toISOString();
-    },
-    parseStringRepr: (valueString: string) => {
-      return new Date(valueString);
-    },
-  },
-  {
-    name: alertPropertyTypes.NAME_FILTER,
-    label: "Name",
-    type: filterTypes.INPUT_TEXT,
-  },
-  {
-    name: alertPropertyTypes.OBSERVABLE_FILTER,
-    label: "Observable",
-    type: filterTypes.CATEGORIZED_VALUE,
-    store: useObservableTypeStore,
-    stringRepr: (filter: { category: observableTypeRead; value: string }) => {
-      return `${filter.category.value}|${filter.value}`;
-    },
-    parseStringRepr: (valueString: string) => {
-      const [category, value] = valueString.split("|");
-      return { category: category, value: value };
-    },
-  },
-  {
-    name: alertPropertyTypes.OBSERVABLE_TYPES_FILTER,
-    label: "Observable Types",
-    type: filterTypes.MULTISELECT,
-    store: useObservableTypeStore,
-    stringRepr: (filter: observableTypeRead[]) => {
-      return filter
-        .map(function (elem) {
-          return elem.value;
-        })
-        .join();
-    },
-    parseStringRepr: (valueString: string) => {
-      return valueString.split(",");
-    },
-  },
-  {
-    name: alertPropertyTypes.OBSERVABLE_VALUE_FILTER,
-    label: "Observable Value",
-    type: filterTypes.INPUT_TEXT,
-  },
-  {
-    name: alertPropertyTypes.OWNER_FILTER,
-    label: "Owner",
-    type: filterTypes.SELECT,
-    store: useUserStore,
-    optionProperty: "displayName",
-    valueProperty: "username",
-  },
-  {
-    name: alertPropertyTypes.QUEUE_FILTER,
-    label: "Queue",
-    type: filterTypes.SELECT,
-    store: useAlertQueueStore,
-    optionProperty: "value",
-    valueProperty: "value",
-  },
-  {
-    name: alertPropertyTypes.TAGS_FILTER,
-    label: "Tags",
-    type: filterTypes.CHIPS,
-    store: useNodeTagStore,
-    stringRepr: (filter: string[]) => {
-      return filter
-        .map(function (elem) {
-          return elem;
-        })
-        .join();
-    },
-    parseStringRepr: (valueString: string) => {
-      return valueString.split(",");
-    },
-  },
-  {
-    name: alertPropertyTypes.THREAT_ACTOR_FILTER,
-    label: "Threat Actor",
-    type: filterTypes.SELECT,
-    store: useNodeThreatActorStore,
-    optionProperty: "value",
-    valueProperty: "value",
-  },
-  {
-    name: alertPropertyTypes.THREATS_FILTER,
-    label: "Threats",
-    type: filterTypes.MULTISELECT,
-    store: useNodeThreatStore,
-    stringRepr: (filter: nodeThreatRead[]) => {
-      return filter
-        .map(function (elem) {
-          return elem.value;
-        })
-        .join();
-    },
-    parseStringRepr: (valueString: string) => {
-      return valueString.split(",");
-    },
-  },
-  {
-    name: alertPropertyTypes.TOOL_FILTER,
-    label: "Tool",
-    type: filterTypes.SELECT,
-    store: useAlertToolStore,
-    optionProperty: "value",
-    valueProperty: "value",
-  },
-  {
-    name: alertPropertyTypes.TOOL_INSTANCE_FILTER,
-    label: "Tool Instance",
-    type: filterTypes.SELECT,
-    store: useAlertToolInstanceStore,
-    optionProperty: "value",
-    valueProperty: "value",
-  },
-  {
-    name: alertPropertyTypes.TYPE_FILTER,
-    label: "Type",
-    type: filterTypes.SELECT,
-    store: useAlertTypeStore,
-    optionProperty: "value",
-    valueProperty: "value",
-  },
-] as const;
-
-export const alertRangeFilters = {
-  "Event Time": {
-    start: alertPropertyTypes.EVENT_TIME_AFTER_FILTER,
-    end: alertPropertyTypes.EVENT_TIME_BEFORE_FILTER,
-  },
-  "Insert Time": {
-    start: alertPropertyTypes.INSERT_TIME_AFTER_FILTER,
-    end: alertPropertyTypes.INSERT_TIME_BEFORE_FILTER,
-  },
-  "Disposition Time": {
-    start: alertPropertyTypes.DISPOSITIONED_AFTER_FILTER,
-    end: alertPropertyTypes.DISPOSITIONED_BEFORE_FILTER,
-  },
-};
+import { inputTypes } from "./base";
 
 // ** Events ** //
 
@@ -310,11 +50,10 @@ export const eventPropertyTypes: Record<string, string> = {
   TYPE_PROPERTY: "type",
   VECTORS_PROPERTY: "vectors",
 };
-
 const eventEventTimeProperty = {
   name: eventPropertyTypes.EVENT_TIME_PROPERTY,
   label: "Event TIme",
-  type: filterTypes.DATE,
+  type: inputTypes.DATE,
   stringRepr: (value: Date): string => {
     return value.toISOString();
   },
@@ -325,7 +64,7 @@ const eventEventTimeProperty = {
 const eventAlertTimeProperty = {
   name: eventPropertyTypes.ALERT_TIME_PROPERTY,
   label: "Alert Time",
-  type: filterTypes.DATE,
+  type: inputTypes.DATE,
   stringRepr: (value: Date): string => {
     return value.toISOString();
   },
@@ -336,7 +75,7 @@ const eventAlertTimeProperty = {
 const eventOwnershipTimeProperty = {
   name: eventPropertyTypes.OWNERSHIP_TIME_PROPERTY,
   label: "Ownership Time",
-  type: filterTypes.DATE,
+  type: inputTypes.DATE,
   stringRepr: (value: Date): string => {
     return value.toISOString();
   },
@@ -347,7 +86,7 @@ const eventOwnershipTimeProperty = {
 const eventDispositionTimeProperty = {
   name: eventPropertyTypes.DISPOSITION_TIME_PROPERTY,
   label: "Disposition Time",
-  type: filterTypes.DATE,
+  type: inputTypes.DATE,
   stringRepr: (value: Date): string => {
     return value.toISOString();
   },
@@ -358,7 +97,7 @@ const eventDispositionTimeProperty = {
 const eventContainTimeProperty = {
   name: eventPropertyTypes.CONTAIN_TIME_PROPERTY,
   label: "Contain Time",
-  type: filterTypes.DATE,
+  type: inputTypes.DATE,
   stringRepr: (value: Date): string => {
     return value.toISOString();
   },
@@ -369,7 +108,7 @@ const eventContainTimeProperty = {
 const eventRemediationTimeProperty = {
   name: eventPropertyTypes.REMEDIATION_TIME_PROPERTY,
   label: "Remediation Time",
-  type: filterTypes.DATE,
+  type: inputTypes.DATE,
   stringRepr: (value: Date): string => {
     return value.toISOString();
   },
@@ -377,11 +116,10 @@ const eventRemediationTimeProperty = {
     return new Date(valueString);
   },
 };
-
 const eventRemediationProperty = {
   name: eventPropertyTypes.REMEDIATIONS_PROPERTY,
   label: "Remediation",
-  type: filterTypes.MULTISELECT,
+  type: inputTypes.MULTISELECT,
   store: useEventRemediationStore,
   stringRepr: (value: eventRemediationRead[]): string => {
     return value
@@ -394,32 +132,28 @@ const eventRemediationProperty = {
     return valueString.split(",");
   },
 };
-
 const eventNameProperty: filterOption = {
   name: eventPropertyTypes.NAME_PROPERTY,
   label: "Name",
-  type: filterTypes.INPUT_TEXT,
+  type: inputTypes.INPUT_TEXT,
 };
-
 const eventOwnerProperty: filterOption = {
   name: eventPropertyTypes.OWNER_PROPERTY,
   label: "Owner",
-  type: filterTypes.SELECT,
+  type: inputTypes.SELECT,
   store: useUserStore,
   optionProperty: "displayName",
   valueProperty: "username",
 };
-
 const eventCommentProperty: filterOption = {
   name: eventPropertyTypes.COMMENTS_PROPERTY,
   label: "Comment",
-  type: filterTypes.INPUT_TEXT,
+  type: inputTypes.INPUT_TEXT,
 };
-
 const eventPreventionToolsProperty: filterOption = {
   name: eventPropertyTypes.PREVENTION_TOOLS_PROPERTY,
   label: "Prevention Tools",
-  type: filterTypes.MULTISELECT,
+  type: inputTypes.MULTISELECT,
   store: useEventPreventionToolStore,
   stringRepr: (value: eventPreventionToolRead[]) => {
     return value
@@ -435,16 +169,15 @@ const eventPreventionToolsProperty: filterOption = {
 const eventRiskLevelProperty: filterOption = {
   name: eventPropertyTypes.RISK_LEVEL_PROPERTY,
   label: "Risk Level",
-  type: filterTypes.SELECT,
+  type: inputTypes.SELECT,
   store: useEventRiskLevelStore,
   optionProperty: "value",
   valueProperty: "value",
 };
-
 const eventStatusProperty: filterOption = {
   name: eventPropertyTypes.STATUS_PROPERTY,
   label: "Status",
-  type: filterTypes.SELECT,
+  type: inputTypes.SELECT,
   store: useEventStatusStore,
   optionProperty: "value",
   valueProperty: "value",
@@ -452,7 +185,7 @@ const eventStatusProperty: filterOption = {
 const eventThreatActorsProperty: filterOption = {
   name: eventPropertyTypes.THREAT_ACTORS_PROPERTY,
   label: "Threat Actors",
-  type: filterTypes.MULTISELECT,
+  type: inputTypes.MULTISELECT,
   store: useNodeThreatActorStore,
   stringRepr: (value: nodeThreatActorRead[]) => {
     return value
@@ -468,7 +201,7 @@ const eventThreatActorsProperty: filterOption = {
 const eventThreatsProperty: filterOption = {
   name: eventPropertyTypes.THREATS_PROPERTY,
   label: "Threats",
-  type: filterTypes.MULTISELECT,
+  type: inputTypes.MULTISELECT,
   store: useNodeThreatStore,
   stringRepr: (value: nodeThreatRead[]) => {
     return value
@@ -481,11 +214,10 @@ const eventThreatsProperty: filterOption = {
     return valueString.split(",");
   },
 };
-
 const eventTypeProperty: filterOption = {
   name: eventPropertyTypes.TYPE_PROPERTY,
   label: "Type",
-  type: filterTypes.SELECT,
+  type: inputTypes.SELECT,
   store: useEventTypeStore,
   optionProperty: "value",
   valueProperty: "value",
@@ -493,7 +225,7 @@ const eventTypeProperty: filterOption = {
 const eventVectorsProperty: filterOption = {
   name: eventPropertyTypes.VECTORS_PROPERTY,
   label: "Vectors",
-  type: filterTypes.MULTISELECT,
+  type: inputTypes.MULTISELECT,
   store: useEventVectorStore,
   stringRepr: (value: eventVectorRead[]) => {
     return value
@@ -531,7 +263,7 @@ export const eventFilters: readonly filterOption[] = [
   {
     name: eventPropertyTypes.CREATED_AFTER_PROPERTY,
     label: "Created After",
-    type: filterTypes.DATE,
+    type: inputTypes.DATE,
     stringRepr: (value: Date) => {
       return value.toISOString();
     },
@@ -542,7 +274,7 @@ export const eventFilters: readonly filterOption[] = [
   {
     name: eventPropertyTypes.CREATED_BEFORE_PROPERTY,
     label: "Created Before",
-    type: filterTypes.DATE,
+    type: inputTypes.DATE,
     stringRepr: (value: Date) => {
       return value.toISOString();
     },
@@ -553,7 +285,7 @@ export const eventFilters: readonly filterOption[] = [
   {
     name: eventPropertyTypes.DISPOSITION_PROPERTY,
     label: "Disposition",
-    type: filterTypes.SELECT,
+    type: inputTypes.SELECT,
     store: useAlertDispositionStore,
     optionProperty: "value",
     valueProperty: "value",
@@ -562,7 +294,7 @@ export const eventFilters: readonly filterOption[] = [
   {
     name: eventPropertyTypes.OBSERVABLE_PROPERTY,
     label: "Observable",
-    type: filterTypes.CATEGORIZED_VALUE,
+    type: inputTypes.CATEGORIZED_VALUE,
     store: useObservableTypeStore,
     stringRepr: (value: { category: observableTypeRead; value: string }) => {
       return `${value.category.value}|${value.value}`;
@@ -575,7 +307,7 @@ export const eventFilters: readonly filterOption[] = [
   {
     name: eventPropertyTypes.OBSERVABLE_TYPES_PROPERTY,
     label: "Observable Types",
-    type: filterTypes.MULTISELECT,
+    type: inputTypes.MULTISELECT,
     store: useObservableTypeStore,
     stringRepr: (value: observableTypeRead[]) => {
       return value
@@ -591,14 +323,14 @@ export const eventFilters: readonly filterOption[] = [
   {
     name: eventPropertyTypes.OBSERVABLE_VALUE_PROPERTY,
     label: "Observable Value",
-    type: filterTypes.INPUT_TEXT,
+    type: inputTypes.INPUT_TEXT,
   },
   eventOwnerProperty,
   eventPreventionToolsProperty,
   {
     name: eventPropertyTypes.QUEUE_PROPERTY,
     label: "Queue",
-    type: filterTypes.SELECT,
+    type: inputTypes.SELECT,
     store: useEventQueueStore,
     optionProperty: "value",
     valueProperty: "value",
@@ -607,7 +339,7 @@ export const eventFilters: readonly filterOption[] = [
   {
     name: eventPropertyTypes.SOURCE_PROPERTY,
     label: "Source",
-    type: filterTypes.SELECT,
+    type: inputTypes.SELECT,
     store: useEventSourceStore,
     optionProperty: "value",
     valueProperty: "value",
@@ -616,7 +348,7 @@ export const eventFilters: readonly filterOption[] = [
   {
     name: eventPropertyTypes.TAGS_PROPERTY,
     label: "Tags",
-    type: filterTypes.CHIPS,
+    type: inputTypes.CHIPS,
     store: useNodeTagStore,
     stringRepr: (value: string[]) => {
       return value
