+++ frontend/tests/unit/src/etc/helpers.spec.ts
@@ -1,4 +1,4 @@
-import { alertFilters } from "@/etc/constants";
+import { alertFilters } from "@/etc/constants/alerts";
 import {
   formatNodeFiltersForAPI,
   parseFilters,
